socket tyding up, slight refactoring and much type and general cleaning up

+++ src/utils.c
@@ -26,7 +26,8 @@ void TIRCFormatError(int lasterror) {
     wchar_t *s = NULL;
     FormatMessageW(
         FORMAT_MESSAGE_ALLOCATE_BUFFER | FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS, 
-        NULL, lasterror,
+        NULL,
+        (DWORD)lasterror,
         MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),
         (LPWSTR)&s, 0, NULL
     );
